requirements.txt
streamlit>=1.28.0
requests>=2.31.0
pillow>=10.0.0
pytesseract>=0.3.10
python-dotenv>=1.0.0
